+++ protected/humhub/modules/content/widgets/views/permaLink.php
@@ -2,7 +2,7 @@
 /* @var $this humhub\components\View */
 ?>
 <li>
-    <a href="#" data-action-click="permalink" data-content-permalink="<?= $permaLink ?>">
+    <a href="#" data-action-click="content.permalink" data-content-permalink="<?= $permaLink ?>">
         <i class="fa fa-link"></i><?= Yii::t('ContentModule.widgets_views_permaLink', 'Permalink') ?>
     </a>
 </li>